Problem 25
Added two new functions to the helpers.py library:
  - fibonacci_r(n): This function returns the nth number in the Fibonacci sequence using the rounding method described on Wikipedia.
  - num_digits(n): This function returns the number of digits of some number n using the decimal library to allow for arbitrary sized n.

Solved the problem itself by starting with n as 100000 (known to result in a very large value for F(n)). We then divide n in half until the length of F(n) is less than 1000 digits. We then increase n until it is exactly 1000 digits.